refactor(server): extract import functions
